Flesh out a little bit more
I want an HTTP response value to encapsulate everything about the
response, such as the full list of headers.  The current HTTP libraries
only give access to specific HTTP headers which are known before the
request.  I need to submit patches to the upstream “http” package to
add low-level support on which I can build.